+++ cnxarchive/tests/test_views.py
@@ -467,198 +467,6 @@ class ViewsTestCase(unittest.TestCase):
         self.fixture.tearDown()
 
 
-    def test_legacy_id_redirect(self):
-        uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, latest version
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(cm.exception.headers['Location'],
-                         quote('/contents/{}@5'.format(uuid)))
-
-    def test_legacy_id_ver_redirect(self):
-        uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid, 'objver': '1.5'}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, latest version
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(cm.exception.headers['Location'],
-                         quote('/contents/{}@5'.format(uuid)))
-
-    # https://github.com/Connexions/cnx-archive/issues/452
-    @testing.db_connect
-    def test_legacy_id_old_ver_collection_with_missing_tree(self, cursor):
-        book_uuid = 'a733d0d2-de9b-43f9-8aa9-f0895036899e'
-        page_uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-        colid = 'col15533'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid, 'objver': '1.4'}
-        self.request.params = {'collection': '{}/latest'.format(colid)}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Delete tree associated with this book.
-        cursor.execute('DELETE FROM trees WHERE documentid = (SELECT module_ident FROM modules WHERE uuid = %s);', (book_uuid,))
-        cursor.connection.commit()
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, old version
-        with self.assertRaises(httpexceptions.HTTPNotFound) as cm:
-            redirect_legacy_content(self.request)
-
-    def test_legacy_id_old_ver_redirect(self):
-        uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid, 'objver': '1.4'}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, old version
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(cm.exception.headers['Location'],
-                         quote('/contents/{}@4'.format(uuid)))
-
-    def test_legacy_bad_id_redirect(self):
-        objid = 'foobar'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, old version
-        with self.assertRaises(httpexceptions.HTTPNotFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '404 Not Found')
-
-    def test_legacy_id_old_ver_collection_context(self):
-        book_uuid = 'a733d0d2-de9b-43f9-8aa9-f0895036899e'
-        page_uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-        colid = 'col15533'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid, 'objver': '1.4'}
-        self.request.params = {'collection': '{}/latest'.format(colid)}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, old version
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(
-            cm.exception.headers['Location'],
-            quote('/contents/{}@1.1:{}'.format(book_uuid, page_uuid)))
-
-    def test_legacy_id_old_ver_bad_collection_context(self):
-        uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-        objid = 'm42709'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid, 'objver': '1.4'}
-        self.request.params = {'collection': 'col45555/latest'}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the new url, old version
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(cm.exception.headers['Location'],
-                         quote('/contents/{}@4'.format(uuid)))
-
-    def test_legacy_filename_redirect(self):
-        uuid = '56f1c5c1-4014-450d-a477-2121e276beca'
-        objid = 'm42081'
-        objver = '1.8'
-        filename = 'Figure_06_03_10a.jpg'
-        sha1 = '95430b74a5ee9e09037c589feb0685ee226a06b8'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid,
-                                  'objver': objver,
-                                  'filename': filename}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view redirects to the resources url
-        with self.assertRaises(httpexceptions.HTTPFound) as cm:
-            redirect_legacy_content(self.request)
-
-        self.assertEqual(cm.exception.status, '302 Found')
-        self.assertEqual(cm.exception.headers['Location'],
-                         quote('/resources/{}/{}'.format(sha1, filename)))
-
-    def test_legacy_no_such_filename_redirect(self):
-        uuid = '56f1c5c1-4014-450d-a477-2121e276beca'
-        objid = 'm42081'
-        objver = '1.8'
-        filename = 'nothere.png'
-
-        # Build the request environment.
-        self.request.matchdict = {'objid': objid,
-                                  'objver': objver,
-                                  'filename': filename}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'legacy-redirect-w-version'
-
-        # Call the view.
-        from ..views import redirect_legacy_content
-
-        # Check that the view 404s
-        self.assertRaises(httpexceptions.HTTPNotFound,
-                          redirect_legacy_content, self.request)
-
     @testing.db_connect
     def test_content_index_html(self, cursor):
         uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
