+++ cnxarchive/views_folder/in_book_search.py
@@ -52,19 +52,6 @@ logger = logging.getLogger('cnxarchive')
 #   Helper functions   #
 # #################### #
 
-def _get_page_in_book(page_uuid, page_version, book_uuid,
-                      book_version, latest=False):
-    book_ident_hash = join_ident_hash(book_uuid, book_version)
-    coltree = _get_content_json(ident_hash=book_ident_hash)['tree']
-    if coltree is None:
-        raise httpexceptions.HTTPNotFound()
-    pages = list(flatten_tree_to_ident_hashes(coltree))
-    page_ident_hash = join_ident_hash(page_uuid, page_version)
-    if page_ident_hash in pages:
-        return book_uuid, '{}:{}'.format(
-            latest and book_uuid or book_ident_hash, page_uuid)
-    # book not in page
-    return page_uuid, page_ident_hash
 
 
 # ######### #
